CRUD complete

+++ ORM-Dapper/Data/IProductRepository.cs
@@ -7,7 +7,7 @@ namespace ORM_Dapper.Data;
 		public IEnumerable<Product> GetAllProducts();
 
 	public void UpdateProduct(int productId, string newName, double newPrice, int categoryId, bool onSale, int stockLevel);
-	public void CreateProduct(string name, int price, int categoryId, bool onSale, int stockLevel);
+	public void CreateProduct(string name, double price, int categoryId, bool onSale, int stockLevel);
 	public void DeleteProduct(int productId);
